Alteração no modelo, e outros arquivos
