+++ exemploEntrada.json
@@ -1,8 +1,17 @@
 {
-    "IFCE-ID":{
-        "ID-ensino":[0, 1, 2, 3],
-        "ID-adm":[0, 1, 2, 3],
-        "ID-galpao":[0, 1, 2, 3],
-        "ID-restauranteacademico":[0, 1, 2, 3]
+    "Id_Instituição":{
+        "id_sensor01":{ 
+            "11/11/2024": 0.42313,
+            "12/11/2024": 0.13343,
+            "13/11/2024": 0.45234,
+            "14/11/2024": 0.21452,
+            "15/11/2024": 0.12352,
+            "16/11/2024": 0.51232,
+            "17/11/2024": ""
+        },
+        "id_sensor02": {
+            "11/11/2024": "",
+            "12/11/2024": 0.34215
+        }
     }
 }
